+++ site/site.config.ts
@@ -1,5 +1,6 @@
 import * as v from "valibot";
 import type { HSvgBrandsIconName } from "zephblaze/core";
+import { defaultSiteConfig } from "zephblaze/core";
 
 export const site = {
     lang: "ja",
@@ -28,3 +29,11 @@ export const postFmSchema = v.object({
 });
 
 export type PostFm = v.InferInput<typeof postFmSchema>;
+
+export default defaultSiteConfig({
+    color: {
+        category: {
+            text: [0x80, 0x6f, 0x61],
+        },
+    },
+});