+++ src/cli/config.ts
@@ -1,7 +1,8 @@
 import { existsSync } from "node:fs";
+import { createRequire } from "node:module";
 import path from "node:path";
 import { cwd } from "node:process";
-import { type DesignRule, default_design_rule } from "@/lib/core/design";
+import { cloneAndMergeRecord } from "@/lib/core/util";
 
 export type AssetConfig = {
     target_prefix: string;
@@ -16,16 +17,16 @@ export type ZephblazeConfig = {
     input: {
         page_dir: string;
         public_dir: string;
+        site_conf: string;
     };
     output: {
         clean_befor_build: boolean;
         dist_dir: string;
     };
     asset: AssetConfig;
-    designrule: DesignRule;
 };
 
-const defaultConfig: ZephblazeConfig = {
+export const default_config: ZephblazeConfig = {
     server: {
         hostname: "localhost",
         port: 4132,
@@ -34,6 +35,7 @@ const defaultConfig: ZephblazeConfig = {
     input: {
         page_dir: "site/pages",
         public_dir: "site/public",
+        site_conf: "site/site.config.ts",
     },
     output: {
         clean_befor_build: true,
@@ -42,31 +44,38 @@ const defaultConfig: ZephblazeConfig = {
     asset: {
         target_prefix: "/assets",
     },
-    designrule: default_design_rule,
 };
 
-export function defineConfig(config: Partial<ZephblazeConfig>): ZephblazeConfig {
-    return { ...config, ...defaultConfig };
+export function loadConfig<T extends Record<string, string | number | symbol | unknown>>(
+    relative_path: string,
+    default_conf: T,
+): T {
+    const require = createRequire(import.meta.url);
+    return requireConfig(require, relative_path, default_conf);
 }
 
-export async function loadConfig(relative_path?: string): Promise<ZephblazeConfig> {
-    const abs_path = path.join(cwd(), relative_path || "zephblaze.config.ts");
+export function requireConfig<T extends Record<string, string | number | symbol | unknown>>(
+    require: NodeJS.Require,
+    relative_path: string,
+    default_conf: T,
+): T {
+    const abs_path = path.join(cwd(), relative_path);
     if (!existsSync(abs_path)) {
         if (relative_path !== undefined) {
-            console.warn(`zephblaze: config file "${abs_path}" is not found. use default.`);
+            console.warn(`zephblaze: config file "${abs_path}" is not found. use default value.`);
         }
-        return defaultConfig;
+        return default_conf;
     }
 
     try {
-        const config = await import(abs_path);
+        const config = require(abs_path);
         if (typeof config.default !== "object") {
             console.warn(`zephblaze: config file "${abs_path}" has no default export. use default configuration.`);
-            return defaultConfig;
+            return default_conf;
         }
-        return config.default;
+        return cloneAndMergeRecord(default_conf, config.default);
     } catch (_e) {
         console.warn(`zephblaze: fail to read config file "${abs_path}". use default.`);
-        return defaultConfig;
+        return default_conf;
     }
 }