+++ CHANGELOG.md
@@ -12,6 +12,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Security: in case of vulnerabilities.
 
 ## [Unreleased]
+## [0.1.0-alpha.7] - 2025-09-27
+### Added
+- import default export of site/site.config.ts to configure original design.
+
+### Changed
+- update error page to show more information.
+
 ## [0.1.0-alpha.6] - 2025-09-26
 ### Fixed
 - fix zephblaze install scheme (with typescript tsc)
@@ -65,7 +72,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Component-level asset management with tree shaking
 - Unified development experience for server-side and client-side
 
-[Unreleased]: https://github.com/osawa-naotaka/zephblaze/compare/v0.1.0-alpha.6...HEAD
+[Unreleased]: https://github.com/osawa-naotaka/zephblaze/compare/v0.1.0-alpha.7...HEAD
+[0.1.0-alpha.7]: https://github.com/osawa-naotaka/zephblaze/compare/v0.1.0-alpha.6...v0.1.0-alpha.7
 [0.1.0-alpha.6]: https://github.com/osawa-naotaka/zephblaze/compare/v0.1.0-alpha.5...v0.1.0-alpha.6
 [0.1.0-alpha.5]: https://github.com/osawa-naotaka/zephblaze/compare/v0.1.0-alpha.4...v0.1.0-alpha.5
 [0.1.0-alpha.4]: https://github.com/osawa-naotaka/zephblaze/compare/v0.1.0-alpha.3...v0.1.0-alpha.4